% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRIO_tools.R
\name{aggregate_Y}
\alias{aggregate_Y}
\title{Aggregates the Y matrix for specfic columns. 
e.g. by country, final demand category}
\usage{
aggregate_Y(Y, groupings)
}
\arguments{
\item{Y}{the final demand matrix}

\item{groupings}{a vector with the groupings, same length as ncol(Y)}
}
\description{
Aggregates the Y matrix for specfic columns. 
e.g. by country, final demand category
}
