% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRIO_tools.R
\name{as.dense.matrix}
\alias{as.dense.matrix}
\title{Title}
\usage{
as.dense.matrix(x, row = 1, col = 2, value = 3, keep.names = TRUE)
}
\arguments{
\item{x}{a sparse matrix in the form a either a data.frame or data.table. Needs to have 3 columns. Default order: row | col | value. If order differs please specify `row`, `col` and `value` arguments. Row and col columns can be either integers (representing the location in the matrix) or character.}

\item{row}{which column of x represent the row-index? default 1}

\item{col}{which column of x represent the column-index? default 2}

\item{value}{which column of x represent the value? default 3}

\item{keep.names}{only considered if the `row` and `col` columns of `x` are of type character.}
}
\value{

}
\description{
Title
}
