% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRIO_tools.R
\name{as.sparse.matrix}
\alias{as.sparse.matrix}
\title{Title}
\usage{
as.sparse.matrix(mat, rownames = NULL, colnames = NULL,
  suffices = c(".row", ".col"))
}
\arguments{
\item{mat}{}
}
\description{
Title
}
