% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_tools.R
\name{big_head}
\alias{big_head}
\title{Returns the upper left corner of a large matrix. Extends the utils::head.matrix 
function by not only return the first n rows but also the first m cols.}
\usage{
big_head(x, n = 6L, m = 10L, ...)
}
\arguments{
\item{x}{a matrix or data.frame}

\item{n}{number of rows to return}

\item{m}{number of cols to return}

\item{...}{}
}
\value{

}
\description{
Returns the upper left corner of a large matrix. Extends the utils::head.matrix 
function by not only return the first n rows but also the first m cols.
}
