% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRIO_tools.R
\name{calc_footprint_sector}
\alias{calc_footprint_sector}
\title{Title}
\usage{
calc_footprint_sector(L_mat, S_mat, y_vec, index, detailed = FALSE)
}
\arguments{
\item{L_mat}{}

\item{S_mat}{}

\item{y_vec}{}

\item{index}{the index of the sector footprints are to be calculated. If missing results for ALL sectors are returned (higher computational expenses)}

\item{detailed}{shall footprints be returned split up by direct + indirect emissions?}
}
\value{

}
\description{
Title
}
