% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_tools.R
\name{clean_session}
\alias{clean_session}
\title{Title}
\usage{
clean_session(detach_packages = TRUE, restart = FALSE, gc = TRUE)
}
\arguments{
\item{restart}{}

\item{gc}{}
}
\value{

}
\description{
Title
}
\examples{

}
