% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_tools.R
\name{coef_var}
\alias{coef_var}
\title{Calculates the coefficient of variance}
\usage{
coef_var(x, na.rm = FALSE)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
Calculates the coefficient of variance
}
