% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_tools.R
\name{combine_words_regexp}
\alias{combine_words_regexp}
\title{If you want to search for two (or more) words within a string in any
possible order}
\usage{
combine_words_regexp(words)
}
\arguments{
\item{words}{a character vector containing the words you want to look for}
}
\value{
an reg exp string
}
\description{
If you want to search for two (or more) words within a string in any
possible order
}
