% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_tools.R
\name{common_elements}
\alias{common_elements}
\title{Title}
\usage{
common_elements(x, y, mode = 1, ignore.case = FALSE)
}
\arguments{
\item{x}{vector}

\item{y}{vector}

\item{mode}{either 1: looking for common elements (default), or -1: looking for non-common elements}
}
\description{
Title
}
