% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eurostat_tools.R
\name{convert_NACErev2_to_tree}
\alias{convert_NACErev2_to_tree}
\title{Converts a data.frame with NACE rev2 classification to a hierarchical data.tree. 
Currently only supports until 2nd level (e.g. A03, C11, ...)}
\usage{
convert_NACErev2_to_tree(x, code)
}
\arguments{
\item{x}{a data.frame}

\item{code}{character: column name with NACE rev2 codes (need to be in the form A, B, ..., A01, C13, C10-12, C10_C11, ...)}
}
\value{
a data.tree object with all columns of x as attributes
}
\description{
Converts a data.frame with NACE rev2 classification to a hierarchical data.tree. 
Currently only supports until 2nd level (e.g. A03, C11, ...)
}
