% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_tools.R
\name{convert_raster2dt}
\alias{convert_raster2dt}
\title{Title}
\usage{
convert_raster2dt(raster, na.rm = TRUE, xy = TRUE)
}
\arguments{
\item{raster}{a raster object}

\item{na.rm}{boolean, should cells with NA values be removed? Default is TRUE}

\item{xy}{boolean, should x- and y- coordinates be listed (as separated columns) in the data.table? Default is TRUE}
}
\value{
a data.table
}
\description{
Title
}
\examples{

}
