% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifications.R
\name{create_concordance_matrix}
\alias{create_concordance_matrix}
\title{Just adds rownames and column names to any matrix. 
rownames with leading i, colnames with leading j}
\usage{
create_concordance_matrix(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
Just adds rownames and column names to any matrix. 
rownames with leading i, colnames with leading j
}
