% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_tools.R
\name{create_named_list}
\alias{create_named_list}
\title{Title}
\usage{
create_named_list(names)
}
\arguments{
\item{names}{a vector (character or numeric). Numeric is converted to character.}
}
\value{
a list of the same length as the names vector
}
\description{
Title
}
\examples{

}
