% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRIO_tools.R
\name{create_random_IOtable}
\alias{create_random_IOtable}
\title{Title}
\usage{
create_random_IOtable(n.industries, n.emissions, n.fdcats, A = FALSE)
}
\arguments{
\item{n.industries}{}

\item{n.emissions}{}

\item{n.fdcats}{}

\item{A}{}
}
\description{
Title
}
