% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRIO_tools.R
\name{.calc.sector.fp.direct}
\alias{.calc.sector.fp.direct}
\title{Title}
\usage{
.calc.sector.fp.direct(S_mat, L_mat, y_vec, index)
}
\arguments{
\item{S_mat}{}

\item{L_mat}{}

\item{y_vec}{}

\item{index}{}
}
\description{
Title
}
