% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRIO_tools.R
\name{.calc.sector.fp.indirect}
\alias{.calc.sector.fp.indirect}
\title{Title}
\usage{
.calc.sector.fp.indirect(S_mat, L_mat, x, index)
}
\arguments{
\item{S_mat}{}

\item{L_mat}{}

\item{x}{}

\item{index}{}
}
\description{
Title
}
