% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eurostat_tools.R
\name{ends_with_2_numbers}
\alias{ends_with_2_numbers}
\title{Checks which elements of a character vector have two digits as their last two characters}
\usage{
ends_with_2_numbers(y)
}
\arguments{
\item{y}{a vector}
}
\value{
vector of booleans
}
\description{
Checks which elements of a character vector have two digits as their last two characters
}
