% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_tools.R
\name{find_a}
\alias{find_a}
\title{Find the "best" a for a log(x+a) transformation when x contains values <= 0
From bestNormalize::log_x}
\usage{
find_a(x, eps = 0.001)
}
\arguments{
\item{x}{}

\item{eps}{}
}
\description{
Find the "best" a for a log(x+a) transformation when x contains values <= 0
From bestNormalize::log_x
}
