% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifications.R
\name{get_most_aggregated_concordances}
\alias{get_most_aggregated_concordances}
\title{Takes two classifications where concordances are known and finds the most 
aggregated classifiaction (i.e. the classification where you do not have to
disaggregate, thus no further knowledge is needed).}
\usage{
get_most_aggregated_concordances(concordance_matrix)
}
\arguments{
\item{concordance_matrix}{}
}
\value{
list with two look-up tables: i2k (concordance of rows to new classifaction), j2k (concordance of cols to new classification)
}
\description{
Takes two classifications where concordances are known and finds the most 
aggregated classifiaction (i.e. the classification where you do not have to
disaggregate, thus no further knowledge is needed).
}
