% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eurostat_tools.R
\name{get_my_eurostat_dic}
\alias{get_my_eurostat_dic}
\title{Get definitions of all Eurostat codes for all variables (=columns) of the data set.}
\usage{
get_my_eurostat_dic(data)
}
\arguments{
\item{data}{A data.frame from get_eurostat(..., type = 'code')}
}
\value{
a list of dictionaries (=data.tables) for each variable of the supplied data (except time and value)
}
\description{
Get definitions of all Eurostat codes for all variables (=columns) of the data set.
}
