% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_tools.R
\name{install_packages}
\alias{install_packages}
\title{Install R packages as RStudio Job}
\usage{
install_packages(pkgs)
}
\arguments{
\item{pkgs}{Package(s) to install}

\item{...}{other arguments passed to pak::pkg_install}
}
\value{

}
\description{
Install R packages as RStudio Job
}
