% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRIO_tools.R
\name{IO_calculator}
\alias{IO_calculator}
\title{Title}
\usage{
IO_calculator(S, L, Y, B, d, f, detailed = TRUE)
}
\arguments{
\item{S}{}

\item{L}{}

\item{Y}{}

\item{B}{}

\item{d}{}

\item{f}{}

\item{detailed}{}
}
\description{
Title
}
