% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eurostat_tools.R
\name{is_coherent_NACErev2}
\alias{is_coherent_NACErev2}
\title{Title}
\usage{
is_coherent_NACErev2(x, code_name = "code", attribute = "value", tol = 0.001)
}
\arguments{
\item{x}{}

\item{code_name}{}

\item{attribute}{}

\item{tol}{}
}
\value{

}
\description{
Title
}
