% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eurostat_tools.R
\name{is_combination}
\alias{is_combination}
\title{Checks which elements of a character vector are combinations of NACErev2 level-2-sectors, such as: 
C10-12 or C10_C11}
\usage{
is_combination(y)
}
\arguments{
\item{y}{a vector}
}
\value{
vector of booleans
}
\description{
Checks which elements of a character vector are combinations of NACErev2 level-2-sectors, such as: 
C10-12 or C10_C11
}
