% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifications.R
\name{is_concordance_matrix}
\alias{is_concordance_matrix}
\title{Checks if x is of type matrix and has rownames and column names}
\usage{
is_concordance_matrix(x)
}
\arguments{
\item{x}{}
}
\value{
boolean
}
\description{
Checks if x is of type matrix and has rownames and column names
}
