% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_tools.R
\name{is_inverse_function}
\alias{is_inverse_function}
\title{Tests if two function are inverse to each other, thus if f(g(x)) == x applies}
\usage{
is_inverse_function(fun1, fun2)
}
\arguments{
\item{fun1}{}

\item{fun2}{}
}
\value{
boolean
}
\description{
Tests if two function are inverse to each other, thus if f(g(x)) == x applies
}
