% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eurostat_tools.R
\name{is_NACE_level1}
\alias{is_NACE_level1}
\title{Checks which elements of a character vector are capital letters from A to U}
\usage{
is_NACE_level1(y)
}
\arguments{
\item{y}{a vector}
}
\value{
vector of booleans
}
\description{
Checks which elements of a character vector are capital letters from A to U
}
