% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eurostat_tools.R
\name{is_NACE_level2}
\alias{is_NACE_level2}
\title{Checks which elements of a character vector are of form [Capital Letter][Digit][Digit], such as
A03, C10, U99}
\usage{
is_NACE_level2(y)
}
\arguments{
\item{y}{a vector}
}
\value{
vector of booleans
}
\description{
Checks which elements of a character vector are of form [Capital Letter][Digit][Digit], such as
A03, C10, U99
}
