% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eurostat_tools.R
\name{is_NACErev2_code}
\alias{is_NACErev2_code}
\title{Checks which elements of a character vector are valid NACE rev2 codes. 
Currently includes: 
- Level 1: A, B, ..., U
- Level 2: A01, C10, U99, ...
- Combinations: C10-12, C10_C11}
\usage{
is_NACErev2_code(y)
}
\arguments{
\item{y}{a vector}
}
\value{
vector of booleans
}
\description{
Not included: Level 3 and 4 codes, such as A01.01 or A01.01.11
}
