% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifications.R
\name{is_Nto1_concordance}
\alias{is_Nto1_concordance}
\title{Checks if a given concordance matrix or lookup table contains ONLY n-to-1 
concordances (i.e. only aggregation needed, no disaggregation)}
\usage{
is_Nto1_concordance(x, origin = 1)
}
\arguments{
\item{x}{either concordance matrix (check if is_concordance_matrix(x) delivers TRUE) or a look-up table with 2 columns}

\item{origin}{if x is a concordance-matrix: either 1 (if the rows indicate the original classification), or 2 (if it is the columns). If x is a look-up table: the argument indicates the column number (or name)}
}
\value{
boolean
}
\description{
Checks if a given concordance matrix or lookup table contains ONLY n-to-1 
concordances (i.e. only aggregation needed, no disaggregation)
}
