% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_tools.R
\name{logap}
\alias{logap}
\title{see ?bestNormalize::log_x()}
\usage{
logap(x, a = max(0, -min(x) + eps), eps = 0.001)
}
\arguments{
\item{x}{}

\item{a}{}

\item{eps}{}
}
\description{
see ?bestNormalize::log_x()
}
