% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifications.R
\name{map_to_classification}
\alias{map_to_classification}
\title{Currently only works for N-to-1 concordances}
\usage{
map_to_classification(origin_id, origin_value, origin_etc, concordance)
}
\arguments{
\item{origin_id}{vector of length N with the ids of the original classification}

\item{origin_value}{vector of length N with the values of the original classification}

\item{origin_etc}{vector of length N OR data.frame with nrow = N with additional attributes (such as Year, Country, ...)}

\item{concordance}{either matrix with given col- and row-names (=concordance matrix) or data.frame. Expressing the way how the original classification shall be mapped to the 'new' classification. If it is a matrix the rows represent the original ids, the columns the new ids. If it is data.frame the first row represents the original ids, the second the new ones.}
}
\value{
a data.frame / data.table (currrently)
}
\description{
Currently only works for N-to-1 concordances
}
