% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_tools.R
\name{normalize}
\alias{normalize}
\title{Normalize vector so that the data has a range from 0 to 1}
\usage{
normalize(x, na.rm = TRUE)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
the normalized vector and the sum of the original vector as attribute (to back-transform later)
}
\description{
Normalize vector so that the data has a range from 0 to 1
}
