% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EB3_tools.R
\name{parse_EB3_SUT}
\alias{parse_EB3_SUT}
\title{Title}
\usage{
parse_EB3_SUT(
  path,
  U = TRUE,
  V = TRUE,
  Y = TRUE,
  va = TRUE,
  metadata = FALSE,
  path2meta = NULL
)
}
\arguments{
\item{path}{}

\item{U}{}

\item{V}{}

\item{Y}{}

\item{va}{}

\item{metadata}{}
}
\value{

}
\description{
Title
}
