% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_tools.R
\name{rescale}
\alias{rescale}
\title{Title}
\usage{
rescale(x, scaling_attributes)
}
\arguments{
\item{x}{a numeric vector or one-column matrix. Optionally: with attributes "scaled:center" and "scaled:scale" (as returned by the base function scale)}

\item{scaling_attributes}{optional.}
}
\value{
a numeric vector. (also in case x is given as a 1-col matrix)
}
\description{
Title
}
