% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_table_tools.R
\name{setroworder}
\alias{setroworder}
\title{Set the row order to a specific given new order
copied from: https://stackoverflow.com/questions/37878620/reorder-rows-in-data-table-in-a-specific-order}
\usage{
setroworder(x, neworder)
}
\arguments{
\item{x}{}

\item{neworder}{}
}
\description{
Set the row order to a specific given new order
copied from: https://stackoverflow.com/questions/37878620/reorder-rows-in-data-table-in-a-specific-order
}
