% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_tools.R
\name{slugify}
\alias{slugify}
\title{Native R slugify (with the help of {stringi})}
\usage{
slugify(x, repl = "-", lower = TRUE)
}
\arguments{
\item{x}{string to slugify}

\item{repl}{what to replace spaces with}

\item{lower}{lowercase final output?}
}
\description{
Native R slugify (with the help of {stringi})
}
