% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRIO_tools.R
\name{SPA_footprint_sector}
\alias{SPA_footprint_sector}
\title{Title}
\usage{
SPA_footprint_sector(n = 8, L_mat, A_mat, y_vec, S_mat, index)
}
\arguments{
\item{n}{number of layers. recommendation >= 8}

\item{L_mat}{}

\item{A_mat}{}

\item{y_vec}{}

\item{S_mat}{}

\item{index}{see ?calc_footprint_sector}
}
\value{

}
\description{
Title
}
