% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eurostat_tools.R
\name{split_NACE_combinations}
\alias{split_NACE_combinations}
\title{Disaggregate Correspondence table with combined NACErev2 sectors}
\usage{
split_NACE_combinations(x, col)
}
\arguments{
\item{x}{a correspondence table (data.table in long format, 2 columns only) where either the source or the target classification include NACErev2 codes that are a combination of several categories.}

\item{col}{numeric: where to find the column with NACErev2 combinations}
}
\value{

}
\description{
Disaggregate Correspondence table with combined NACErev2 sectors
}
\details{
Combinations of sectors can either have the form C10_C11 (type underscore)
or C10-C13 (type bar). 

_ implies 'and' (C10_C11 = C10 and C11)
- implies 'from to' (C10-C13 = C10, C11, C12, C13)
}
\examples{

}
