% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwunicluster_tools.R
\name{ssh_install_packages}
\alias{ssh_install_packages}
\title{Title}
\usage{
ssh_install_packages(session, pkgs, ...)
}
\arguments{
\item{session}{}

\item{pkgs}{}

\item{...}{}
}
\value{

}
\description{
Title
}
