% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwunicluster_tools.R
\name{ssh_new_project}
\alias{ssh_new_project}
\title{Title}
\usage{
ssh_new_project(session, project_name, ...)
}
\arguments{
\item{session}{}

\item{project_name}{}

\item{...}{}
}
\value{

}
\description{
Title
}
