% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_management_tools.R
\name{start_new_project}
\alias{start_new_project}
\title{Run this function when starting a new projects. 
1. Creates a new directory with all relevant folders (code, tex, data, figures, ...)
2. Optionally creates an Rproject within the Code folder
3. Optinally with Git/Github}
\usage{
start_new_project(
  name,
  root_dir = "/home/simon/Documents/Projects",
  rproject = name,
  git = TRUE,
  github = TRUE
)
}
\arguments{
\item{name}{the name of the project/the folder supposed to be created}

\item{root_dir}{the root directory where the project should be stored}

\item{rproject}{defaults to `name`, can be given any character string. Set to NULL if you wish not to have a rproject be created}

\item{git}{TRUE if you want to create a local git repo. You will be asked to make an initial commit}

\item{github}{TRUE if you wish to create an remote repo on github}
}
\value{

}
\description{
Run this function when starting a new projects. 
1. Creates a new directory with all relevant folders (code, tex, data, figures, ...)
2. Optionally creates an Rproject within the Code folder
3. Optinally with Git/Github
}
