% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_management_tools.R
\name{start_R_script}
\alias{start_R_script}
\title{Title}
\usage{
start_R_script(
  filename = "Untitled.R",
  dir = ".",
  author = "Simon Schulte",
  packages = TRUE
)
}
\arguments{
\item{filename}{}

\item{dir}{}

\item{author}{}

\item{packages}{}
}
\value{

}
\description{
Title
}
