% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eurostat_tools.R
\name{starts_with_NACE_letter}
\alias{starts_with_NACE_letter}
\title{Checks which elements of a character vector have a capital letter from A to U as their first character}
\usage{
starts_with_NACE_letter(y)
}
\arguments{
\item{y}{a vector}
}
\value{
vector of booleans
}
\description{
Checks which elements of a character vector have a capital letter from A to U as their first character
}
