% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tools.R
\name{stat_boxplot_custom}
\alias{stat_boxplot_custom}
\title{Modifies the ggplot geom_boxplot function to extend the whiskers to a specific
percentile instead of 1.5 QR. 
Copied from: https://gist.github.com/rabutler/bd97a6f49db87860f987156842fd4ee5}
\usage{
stat_boxplot_custom(
  mapping = NULL,
  data = NULL,
  geom = "boxplot",
  position = "dodge",
  ...,
  qs = c(0.05, 0.25, 0.5, 0.75, 0.95),
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{}

\item{data}{}

\item{geom}{}

\item{position}{}

\item{...}{}

\item{qs}{}

\item{na.rm}{}

\item{show.legend}{}

\item{inherit.aes}{}
}
\value{

}
\description{
Modifies the ggplot geom_boxplot function to extend the whiskers to a specific
percentile instead of 1.5 QR. 
Copied from: https://gist.github.com/rabutler/bd97a6f49db87860f987156842fd4ee5
}
