% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_tools.R
\name{substr_right}
\alias{substr_right}
\title{Extracts the last n characters of a character string x}
\usage{
substr_right(x, n)
}
\arguments{
\item{x}{a character string}

\item{n}{the last n elements to extract}
}
\value{
a character string of length n
}
\description{
Extracts the last n characters of a character string x
}
