% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_tools.R
\name{transform}
\alias{transform}
\title{Transform a data. Here 'transform' is used in a wider meaning including standardization and normalization
This function can be used to 
(i) transform `x` using the function given by `fun`
(ii) scale `x` using R's `scale` function (=standardization)
(iii) normalize `x` so that its values have a range from 0 to 1}
\usage{
transform(
  x,
  fun,
  fun_inverse,
  scale = FALSE,
  normalize = FALSE,
  eps = 0.001,
  ...
)
}
\arguments{
\item{x}{a numeric vector}

\item{fun}{a function to transform x. Or character vector: at the moment only 'logap' (log(x+a)) is implemented}

\item{fun_inverse}{the inverse of it}

\item{scale}{boolean. shall x also be scaled? default = FALSE}

\item{normalize}{boolean. shall x also be normalized? default = FALSE}

\item{...}{other parameters passed `scale` (at the moment only support `center`)}
}
\value{

}
\description{
When combining (i) and (ii) or (i) and (iii) the data is first transformed then scaled/normalized
Note: Combining (ii) and (iii) does not make any sense.
}
