% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifications.R
\name{transform_concordance_matrix_to_lookup_table}
\alias{transform_concordance_matrix_to_lookup_table}
\title{Concordance matrix = matrix which has unique row- and colnames
Lookup table = 2-column data.frame/table with row-wise unique concordances}
\usage{
transform_concordance_matrix_to_lookup_table(concordance_matrix)
}
\arguments{
\item{concordance_matrix}{}
}
\value{

}
\description{
Concordance matrix = matrix which has unique row- and colnames
Lookup table = 2-column data.frame/table with row-wise unique concordances
}
