% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwunicluster_tools.R
\name{use_bwunicluster}
\alias{use_bwunicluster}
\title{Title}
\usage{
use_bwunicluster(
  ssh_address = "fr_ss1625@bwunicluster.scc.kit.edu",
  file_to_execute,
  files_etc = NULL,
  project_name,
  n_nodes = NA,
  n_cores = NA,
  time_minutes = NA,
  RAM_per_node = NA,
  ...
)
}
\arguments{
\item{file_to_execute}{}

\item{files_etc}{}

\item{project_name}{}

\item{n_nodes}{}

\item{n_cores}{}

\item{time_minutes}{}

\item{RAM_per_node}{}

\item{...}{}

\item{session}{}
}
\value{

}
\description{
Title
}
