% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_tools.R
\name{weighted.var}
\alias{weighted.var}
\title{Title
from:   https://rdrr.io/github/hadley/bigvis/man/weighted.var.html}
\usage{
weighted.var(x, w = NULL, na.rm = FALSE)
}
\arguments{
\item{x}{}

\item{w}{}

\item{na.rm}{}
}
\value{

}
\description{
Title
from:   https://rdrr.io/github/hadley/bigvis/man/weighted.var.html
}
