% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_tools.R
\name{zero_range}
\alias{zero_range}
\title{Title
see https://www.rdocumentation.org/packages/scales/versions/0.4.1/topics/zero_range}
\usage{
zero_range(x, tol = .Machine$double.eps^0.5, na.rm = TRUE)
}
\arguments{
\item{x}{}

\item{tol}{}
}
\description{
Title
see https://www.rdocumentation.org/packages/scales/versions/0.4.1/topics/zero_range
}
